refactor: useChatRoom hook add (#135)
